refactor(Intl Address): Reorganise for international addresses
Major version bump as a precaution

BREAKING CHANGE:

Major version bump as a precaution

API Paths have been retagged and reorganised. This should not cause a
change in functionality but we have noticed changes of this type can
sometimes produce a backwards incompatible change in API interface if
this package is consumed by a code generator.